Make the internal skin storage resource provider public.
